feat: Parallelize stream processing with workers threads
